+++ server/src/Storage/Entry.h
@@ -32,6 +32,7 @@ namespace Sider::Storage::Entry
     struct Entry
     {
         virtual Type type() const = 0;
+        virtual const std::string get() = 0;
         virtual bool isExpired() = 0;
         virtual ~Entry() = default;
     };
@@ -39,20 +40,12 @@ namespace Sider::Storage::Entry
     struct CounterEntry : public Entry
     {
         virtual void increment(Ttl ttl, uint8_t step) = 0;
-        virtual const uint16_t get() = 0;
-    };
-    
-    struct KeeperEntry : public Entry
-    {
-        virtual void keep(Ttl ttl, std::string value) = 0;
-        virtual const std::string get() = 0;
     };
 
     struct QueueEntry : public Entry
     {
         virtual void queue(std::string value) = 0;
         virtual void dequeue() = 0;
-        virtual const std::string get() = 0;
     };
 
     struct RaterEntry : public Entry
@@ -61,7 +54,6 @@ namespace Sider::Storage::Entry
         virtual const uint16_t get(uint8_t partition) = 0;
     };
 
-    KeeperEntry* initKeeperEntry();
     CounterEntry* initCounterEntry();
     QueueEntry* initQueueEntry();
     RaterEntry* initRaterEntry();
