+++ server/src/Storage/Storage.cpp
@@ -50,6 +50,12 @@ namespace Sider::Storage
             entries[scope.name].clear();
         }
 
+        void add(const Entry::Id id, Entry::Entry *entry) override
+        {
+            // todo do not allow to add entry with same id
+            entries[id.scope.name][id.key.name] = entry;
+        }   
+
         Entry::Entry* find(const Entry::Id id) override
         {
             if (!entries[id.scope.name].contains(id.key.name)) {
@@ -71,10 +77,10 @@ namespace Sider::Storage
             return static_cast<Entry::CounterEntry*>(get(id, Entry::Type::COUNTER));
         }
 
-        Entry::KeeperEntry* getKeeper(const Entry::Id id) override
-        {
-            return static_cast<Entry::KeeperEntry*>(get(id, Entry::Type::KEEPER));
-        }
+        // Entry::KeeperEntry* getKeeper(const Entry::Id id) override
+        // {
+        //     return static_cast<Entry::KeeperEntry*>(get(id, Entry::Type::KEEPER));
+        // }
 
         Entry::QueueEntry* getQueue(const Entry::Id id) override
         {
@@ -102,7 +108,7 @@ namespace Sider::Storage
         {
             switch (type) {
                 case Entry::Type::COUNTER: return Entry::initCounterEntry();
-                case Entry::Type::KEEPER: return Entry::initKeeperEntry();
+                // case Entry::Type::KEEPER: return Entry::initKeeperEntry();
                 case Entry::Type::QUEUE: return Entry::initQueueEntry();
                 case Entry::Type::RATER: return Entry::initRaterEntry();
                 default: throw std::runtime_error("Unsupported entry type '" + id.scope.name + ":" + id.key.name + "'");
