+++ server/src/Storage/Storage.h
@@ -12,9 +12,10 @@ namespace Sider::Storage
         virtual void clear(const Entry::Id id) = 0;
         virtual void clear(const Entry::Scope scope) = 0;
 
+        virtual void add(const Entry::Id id, Entry::Entry *entry) = 0;
+
         virtual Entry::Entry* find(const Entry::Id id) = 0;
         virtual Entry::CounterEntry* getCounter(const Entry::Id id) = 0;
-        virtual Entry::KeeperEntry* getKeeper(const Entry::Id id) = 0;
         virtual Entry::QueueEntry* getQueue(const Entry::Id id) = 0;
         virtual Entry::RaterEntry* getRater(const Entry::Id id) = 0;
     };
